% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_info.R
\name{acc_info}
\alias{acc_info}
\title{Dropbox account information}
\usage{
acc_info()
}
\description{
Retrieves information about the user's account. Returns the following fields:  referral_link, display_name,  uid, locale, email_verified, quota_info.datastores, quota_info.shared, quota_info.quota, quota_info.normal, is_paired, country, name_details.familiar_name, name_details.surname, name_details.given_name, email
}
\examples{
\dontrun{
acc_info()
# Select name and UID
acc_info() \%>\% select(uid, display_name, quota_info.normal)
}
}

