% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_file_ops.R
\name{drop_delete}
\alias{drop_delete}
\title{Deletes a file or folder.}
\usage{
drop_delete(path = NULL, root = "auto")
}
\arguments{
\item{path}{This is required The path to the new folder to create relative to root.}

\item{root}{This is required. The root relative to which path is specified. Valid values are auto (recommended), sandbox, and dropbox.}
}
\description{
Deletes a file or folder.
}

