% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_dir.R
\name{drop_dir}
\alias{drop_dir}
\title{drop_dir}
\usage{
drop_dir(path, file_limit = 1000, hash = NULL, list = TRUE,
  include_deleted = FALSE, rev = NULL, locale = NULL,
  include_media_info = TRUE, include_membership = FALSE)
}
\arguments{
\item{path}{The directory listing relative to the root folder.}

\item{file_limit}{Default is 10000. (max is 25000). When listing a folder, the service won't report listings containing more than the specified amount of files and will instead respond with a 406 (Not Acceptable) status response.}

\item{hash}{Each call to /metadata on a folder will return a hash field, generated by hashing all of the metadata contained in that response. On later calls to /metadata, you should provide that value via this parameter so that if nothing has changed, the response will be a 304 (Not Modified) status code instead of the full, potentially very large, folder listing. This parameter is ignored if the specified path is associated with a file or if list=false.}

\item{list}{The strings true and false are valid values. true is the default. If true, the folder's metadata will include a contents field with a list of metadata entries for the contents of the folder. If false, the contents field will be omitted.}

\item{include_deleted}{Only applicable when list is set. If this parameter is set to true, then contents will include the metadata of deleted children. Note that the target of the metadata call is always returned even when it has been deleted (with is_deleted set to true) regardless of this flag.}

\item{rev}{If you include a particular revision number, then only the metadata for that revision will be returned.}

\item{locale}{Dropbox uses the locale parameter to specify language settings of content responses. If your app supports any language other than English, insert the appropriate IETF language tag. When a supported language is specified, Dropbox will returned translated size and/or user_error fields (where applicable).}

\item{include_media_info}{If true, each file will include a photo_info dictionary for photos and a video_info dictionary for videos with additional media info. If the data isn't available yet, the string pending will be returned instead of a dictionary.}

\item{include_membership}{If true, metadata for a shared folder will include a list of members and a list of groups.}
}
\description{
Retrieves file and folder metadata.
}
\details{
A sample output for contents looks like this in the unparsed state
{
"size": "225.4KB",
"rev": "35e97029684fe",
"thumb_exists": false,
"bytes": 230783,
"modified": "Tue, 19 Jul 2011 21:55:38 +0000",
"client_mtime": "Mon, 18 Jul 2011 18:04:35 +0000",
"path": "/Getting_Started.pdf",
"is_dir": false,
"icon": "page_white_acrobat",
"root": "dropbox",
"mime_type": "application/pdf",
"revision": 220823
}
}
\examples{
\dontrun{
drop_dir()
}
}

