% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_dir.R
\name{drop_dir_internal}
\alias{drop_dir_internal}
\title{Old internal function used by drop_delete.}
\usage{
drop_dir_internal(path = NULL, file_limit = 10000, hash = NULL,
  list = TRUE, include_deleted = FALSE, rev = NULL, locale = NULL,
  include_media_info = TRUE, include_membership = FALSE, verbose = FALSE,
  dtoken = get_dropbox_token())
}
\description{
TODO: update drop_delete to not need this, then delete it.
}
\details{
noRd
}
\keyword{internal}
