% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_shared.R
\name{drop_list_shared_links}
\alias{drop_list_shared_links}
\title{List all shared links}
\usage{
drop_list_shared_links(verbose = TRUE, dtoken = get_dropbox_token())
}
\arguments{
\item{verbose}{Print verbose output}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\description{
This function returns a list of all links that are currently being shared
}
\examples{
\dontrun{
drop_list_shared_links()
}
}
