% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_mime.R
\name{drop_mime}
\alias{drop_mime}
\title{Computes mime type from file extension.}
\usage{
drop_mime(path)
}
\arguments{
\item{path}{path to file}
}
\value{
mime string
}
\description{
Computes mime type from file extension.
}
\keyword{internal}

