% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_file_ops.R
\name{drop_move}
\alias{drop_move}
\title{Moves a file or folder to a new location.}
\usage{
drop_move(root, from_path = NULL, to_path = NULL, verbose = FALSE,
  dtoken = get_dropbox_token())
}
\arguments{
\item{root}{This is required. The root relative to which path is specified. Valid values are auto (recommended), sandbox, and dropbox.}

\item{from_path}{Source file or folder}

\item{to_path}{destination file or folder}

\item{verbose}{By default verbose output is \code{FALSE}. Set to \code{TRUE} if you need to troubleshoot any output or grab additional parameters.}

\item{dtoken}{The Dropbox token}
}
\description{
Moves a file or folder to a new location.
}

