% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_put.R
\name{drop_put}
\alias{drop_put}
\title{Uploads a file to Dropbox using PUT semantics.}
\usage{
drop_put(file, dest = NULL, encode = "multipart")
}
\arguments{
\item{file}{Full local path to file.}

\item{dest}{The relative path on Dropbox where the file should get uploaded.}

\item{encode}{= "multipart" encoding format.}
}
\description{
Uploads a file to Dropbox using PUT semantics.
}
\examples{
\dontrun{
drop_put()
}
}

