% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_read.R
\name{drop_read}
\alias{drop_read}
\title{drop_read}
\usage{
drop_read(file, dest = tempdir(), dtoken = get_dropbox_token(), ...)
}
\arguments{
\item{file}{path on dropbox}

\item{dest}{local path. tempdir for default}

\item{dtoken}{token}

\item{...}{other arguments according to file format
into \code{read.csv} or \code{read_excel} or \code{readRDS} or \code{load}}
}
\description{
wrapper for importing read.csv,
read_excel readRDS and load from dropbox
}
\examples{
\dontrun{
save(airquality, file = "airquality.RData")
save(attenu, file = "attenu.RData")
save(austres, file = "austres.RData")
saveRDS(AirPassengers, "AirPassengers.rds")
write.csv(mtcars, file = "mtcars.csv")
openxlsx::write.xlsx(iris, file = "iris.xlsx")
purrr::walk(c("airquality.RData",
              "attenu.RData",
              "austres.RData",
              "AirPassengers.rds",
              "mtcars.csv",
              "iris.xlsx"),

            ~ rdrop2::drop_upload(.x,
                          path = "/", # path in dropbox
                          mode = "overwrite"
            ))
drop_read(file = "AirPassengers.rds")
drop_read("iris.xlsx")
drop_read("mtcars.csv")
drop_read("airquality.RData")
drop_read("attenu.RDATA")
drop_read("austres.rdata")

}
}
\author{
Lewis Hounkpevi
}
