% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_history.R
\name{drop_restore}
\alias{drop_restore}
\title{Restore a file to a specific revision.}
\usage{
drop_restore(path, rev, dtoken = get_dropbox_token())
}
\arguments{
\item{path}{Path to the file on Dropbox.}

\item{rev}{The revision identifier string (e.g. \code{"a1c10ce0dd78"}) to
restore to.  Revision IDs are returned in the \code{rev} column of
\code{\link{drop_history}}.}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\value{
A list of file metadata reflecting the restored version.
}
\description{
Reverts a file on Dropbox to the content it had at a given revision. Use
\code{\link{drop_history}} to find available revision IDs.
}
\examples{
\dontrun{
  history <- drop_history("report.csv")
  # restore to the second most recent version
  drop_restore("report.csv", rev = history$rev[2])
}
}
\references{
\href{https://www.dropbox.com/developers/documentation/http/documentation#files-restore}{API documentation}
}
