% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_search.R
\name{drop_search_continue}
\alias{drop_search_continue}
\title{Continue a paginated search begun with \code{drop_search}.}
\usage{
drop_search_continue(cursor, dtoken = get_dropbox_token())
}
\arguments{
\item{cursor}{A cursor string returned in a previous \code{drop_search} call.}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\value{
Same structure as \code{\link{drop_search}}.
}
\description{
Continue a paginated search begun with \code{drop_search}.
}
\examples{
\dontrun{
  first_page  <- drop_search("report", max_results = 20)
  second_page <- drop_search_continue(first_page$cursor)
}
}
\references{
\href{https://www.dropbox.com/developers/documentation/http/documentation#files-search-continue_v2}{API documentation}
}
