% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_acc.R
\name{drop_space_usage}
\alias{drop_space_usage}
\title{Get Dropbox storage space usage.}
\usage{
drop_space_usage(dtoken = get_dropbox_token())
}
\arguments{
\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\value{
A list with elements \code{used} (bytes used) and \code{allocation}
  (a list with \code{.tag} and \code{allocated} bytes, or team-level info).
}
\description{
Returns how much space the current account is using and how much is
allocated.
}
\examples{
\dontrun{
  usage <- drop_space_usage()
  cat("Used:", usage$used, "bytes\n")
  cat("Allocated:", usage$allocation$allocated, "bytes\n")
}
}
\references{
\href{https://www.dropbox.com/developers/documentation/http/documentation#users-get_space_usage}{API documentation}
}
