% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_upload.R
\name{drop_upload}
\alias{drop_upload}
\title{Uploads a file to Dropbox.}
\usage{
drop_upload(
  file,
  path = NULL,
  mode = "overwrite",
  autorename = TRUE,
  mute = FALSE,
  verbose = FALSE,
  dtoken = get_dropbox_token()
)
}
\arguments{
\item{file}{Relative path to local file.}

\item{path}{The relative path on Dropbox where the file should get uploaded.}

\item{mode}{Upload mode: \code{"overwrite"} (default) will always overwrite
an existing file; \code{"add"} will not overwrite and instead create a
renamed copy on conflict; \code{"update"} requires a \code{rev} argument.}

\item{autorename}{If \code{TRUE} (default), the file being uploaded will be
automatically renamed to avoid conflicts when using \code{mode = "add"}.}

\item{mute}{Set to \code{TRUE} to suppress desktop/mobile notifications.
Defaults to \code{FALSE}.}

\item{verbose}{By default verbose output is \code{FALSE}. Set to \code{TRUE}
if you need to troubleshoot any output or grab additional parameters.}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\value{
Dropbox file metadata list, invisibly.
}
\description{
This function will allow you to write files of any size to Dropbox. Files
larger than 150 MB are automatically uploaded in chunks using the Dropbox
upload session API.
}
\examples{
\dontrun{
write.csv(mtcars, file = "mtt.csv")
drop_upload("mtt.csv")
}
}
\references{
\href{https://www.dropbox.com/developers/documentation/http/documentation#files-upload}{API documentation}
}
