% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_info.R
\name{drop_acc}
\alias{drop_acc}
\title{Dropbox account information}
\usage{
drop_acc(dtoken = get_dropbox_token(), verbose = FALSE)
}
\arguments{
\item{dtoken}{The Dropbox token generated by \code{\link{dropbox_auth}}. rDrop2 will try to automatically locate your local credential cache and use them. However, if the credentials are not found, the function will initiate a new authentication request. You can override this by pointing to a different location where your credentials are stored.}

\item{verbose}{By default verbose output is \code{FALSE}. Set to \code{TRUE} if you need to troubleshoot any output or grab additional parameters.}
}
\description{
Retrieves information about the user's account. Returns the following fields:  referral_link, display_name,  uid, locale, email_verified, quota_info.datastores, quota_info.shared, quota_info.quota, quota_info.normal, is_paired, country, name_details.familiar_name, name_details.surname, name_details.given_name, email
}
\examples{
\dontrun{
drop_acc()
# Select name and UID
library(dplyr)
drop_acc() \%>\% select(uid, display_name, quota_info.normal)
}
}

