% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acc_info.R
\name{drop_acc}
\alias{drop_acc}
\title{Dropbox account information}
\usage{
drop_acc(dtoken = get_dropbox_token(), verbose = FALSE)
}
\arguments{
\item{dtoken}{The Dropbox token}

\item{verbose}{By default verbose output is \code{FALSE}. Set to \code{TRUE} if you need to troubleshoot any output or grab additional parameters.}
}
\description{
Retrieves information about the user's account. Returns the following fields:  referral_link, display_name,  uid, locale, email_verified, quota_info.datastores, quota_info.shared, quota_info.quota, quota_info.normal, is_paired, country, name_details.familiar_name, name_details.surname, name_details.given_name, email
}
\examples{
\dontrun{
drop_acc()
# Select name and UID
drop_acc() \%>\% select(uid, display_name, quota_info.normal)
}
}

