% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_auth.R
\name{drop_auth}
\alias{drop_auth}
\title{Authentication for Dropbox}
\usage{
drop_auth(new_user = FALSE)
}
\arguments{
\item{new_user}{Default is \code{FALSE}. Set to \code{TRUE} if you need to switch to a new user account.}
}
\description{
This function authenticates you into Dropbox
}
\examples{
\dontrun{
dropbox_auth()
}
}

