% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_file_ops.R
\name{drop_create}
\alias{drop_create}
\title{Creates a folder.}
\usage{
drop_create(path = NULL, root = "auto")
}
\arguments{
\item{path}{This is required The path to the new folder to create relative to root.}

\item{root}{This is required. The root relative to which path is specified. Valid values are auto (recommended), sandbox, and dropbox.}
}
\description{
Returns the following type of response
{
    "size": "0 bytes",
    "rev": "1f477dd351f",
    "thumb_exists": false,
    "bytes": 0,
    "modified": "Wed, 10 Aug 2011 18:21:30 +0000",
    "path": "/new_folder",
    "is_dir": true,
    "icon": "folder",
    "root": "dropbox",
    "revision": 5023410
}
}
\examples{
\dontrun{
drop_create(path = "")
}
}

