% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_file_ops.R
\name{drop_delete}
\alias{drop_delete}
\title{Deletes a file or folder.}
\usage{
drop_delete(path = NULL, verbose = FALSE, dtoken = get_dropbox_token())
}
\arguments{
\item{path}{This is required The path to the new folder to create relative to
root.}

\item{verbose}{By default verbose output is \code{FALSE}. Set to \code{TRUE}
if you need to troubleshoot any output or grab additional parameters.}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\description{
Deletes a file or folder.
}
