% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_dir.R
\name{drop_dir}
\alias{drop_dir}
\title{List folder contents and associated metadata}
\usage{
drop_dir(path = "", recursive = FALSE, include_media_info = FALSE,
  include_deleted = FALSE, include_has_explicit_shared_members = FALSE,
  include_mounted_folders = TRUE, limit = NULL,
  dtoken = get_dropbox_token())
}
\arguments{
\item{path}{path to folder in Dropbox to list contents of. Defaults to the root directory.}

\item{recursive}{If true, the list folder operation will be applied recursively to all subfolders and the response will contain contents of all subfolders. Defaults to FALSE.}

\item{include_media_info}{If true, FileMetadata.media_info is set for photo and video. Defaults to FALSE.}

\item{include_deleted}{If true, the results will include entries for files and folders that used to exist but were deleted. Defaults to FALSE.}

\item{include_has_explicit_shared_members}{If true, the results will include a flag for each file indicating whether or not that file has any explicit members. Defaults to FALSE.}

\item{include_mounted_folders}{If true, the results will include entries under mounted folders which includes app folder, shared folder and team folder. Defaults to TRUE.}

\item{limit}{The maximum number of results to return per request. Note: This is an approximate number and there can be slightly more entries returned in some cases. Defaults to NULL, no limit.}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\value{
\code{tbl_df} of items in folder, one row per file or folder, with metadata values as columns.
}
\description{
List folder contents and associated metadata
}
\examples{
\dontrun{
  drop_dir()
}

}
