% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_get.R
\name{drop_get}
\alias{drop_get}
\title{Downloads a file from Dropbox}
\usage{
drop_get(path = NULL, local_file = NULL, overwrite = FALSE,
  dtoken = get_dropbox_token())
}
\arguments{
\item{path}{This is required The path to the new folder to create relative to
root.}

\item{local_file}{The name of the local copy. Leave this blank if you're fine with the original name.}

\item{overwrite}{Default is \code{FALSE} but can be set to \code{TRUE}.}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}.
dropper will try to automatically locate your local credential cache and use
them. However, if the credentials are not found, the function will initiate
a new authentication request. You can override this by pointing to a
different location where your credentials are stored.}
}
\description{
Downloads a file from Dropbox
}
\examples{
\dontrun{
drop_get(path = 'karthik_small.png', dest = "~/Desktop")
# To overwrite the existing file
 drop_get(path = 'karthik_small.png', overwrite = TRUE)
}
}

