% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_shared.R
\name{drop_get_shared_link_file}
\alias{drop_get_shared_link_file}
\title{List shared link for specific file}
\usage{
drop_get_shared_link_file(verbose = TRUE, dtoken = get_dropbox_token(),
  path)
}
\arguments{
\item{verbose}{Print verbose output}

\item{dtoken}{The Dropbox token generated by \code{\link{drop_auth}}. rdrop2
will try to automatically locate your local credential cache and use them.
However, if the credentials are not found, the function will initiate a new
authentication request. You can override this in \code{\link{drop_auth}} by
pointing to a different location where your credentials are stored.}
}
\description{
List shared link for specific file
}
\examples{
\dontrun{
drop_get_shared_link_file()
}
}
\references{
\href{https://www.dropbox.com/developers/documentation/http/documentation#get_shared_link_file}{API documentation}
}
