% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_mine.R
\name{drop_mime}
\alias{drop_mime}
\title{Compute mime type from file extension.
Needs mechanism for user extension.}
\usage{
drop_mime(path)
}
\arguments{
\item{path}{path to file}
}
\value{
mime string
}
\description{
Compute mime type from file extension.
Needs mechanism for user extension.
}
\keyword{internal}

