% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_search.R
\name{drop_search}
\alias{drop_search}
\title{Returns metadata for all files and folders whose filename contains the given search string as a substring.}
\usage{
drop_search(query = NULL, path = NULL, file_limit = 1000,
  include_deleted = FALSE, locale = NULL, include_membership = FALSE,
  dtoken = get_dropbox_token())
}
\arguments{
\item{query}{}

\item{path}{}

\item{file_limit}{Default is 1000.}

\item{include_deleted}{}

\item{locale}{<what param does>}

\item{include_membership}{}

\item{dtoken}{The Dropbox token}

\item{verbose}{By default verbose output is \code{FALSE}. Set to \code{TRUE} if you need to troubleshoot any output or grab additional parameters.}
}
\description{
Returns metadata for all files and folders whose filename contains the given search string as a substring.
}
\examples{
\dontrun{
drop_search('gif')
}
}

