% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_upload.R
\name{drop_upload}
\alias{drop_upload}
\title{Uploads a file to Dropbox using PUT semantics.}
\usage{
drop_upload(file, dest = NULL, encode = "multipart", verbose = FALSE,
  dtoken = get_dropbox_token())
}
\arguments{
\item{file}{Full local path to file.}

\item{dest}{The relative path on Dropbox where the file should get uploaded.}

\item{encode}{File encoding}

\item{verbose}{default is FALSE}

\item{dtoken}{The Dropbox token generated by \code{\link{dropbox_auth}}.
rDrop2 will try to automatically locate your local credential cache and use
them. However, if the credentials are not found, the function will initiate
a new authentication request. You can override this by pointing to a
different location where your credentials are stored.}
}
\description{
Uploads a file to Dropbox using PUT semantics.
}
\examples{
\dontrun{
write.csv(mtcars, file = "mtcars.csv")
drop_upload("mtcars.csv")
}
}

