% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_put.R
\name{drop_upload}
\alias{drop_upload}
\title{Uploads a file to Dropbox using PUT semantics.}
\usage{
drop_upload(file, dest = NULL, encode = "multipart", verbose = FALSE,
  dtoken = get_dropbox_token())
}
\arguments{
\item{file}{Full local path to file.}

\item{dest}{The relative path on Dropbox where the file should get uploaded.}

\item{encode}{File encoding}

\item{verbose}{default is FALSE}

\item{dtoken}{The Dropbox token}
}
\description{
Uploads a file to Dropbox using PUT semantics.
}
\examples{
\dontrun{
write.csv(mtcars, file = "mtcars.csv")
drop_upload("mtcars.csv")
}
}

