% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_auth.R
\name{get_dropbox_token}
\alias{get_dropbox_token}
\title{Retrieve oauth2 token from rdrop2-namespaced environment}
\usage{
get_dropbox_token()
}
\description{
Retrieves a token if it is previously stored, otherwise prompts user to get one.
}
\keyword{internal}
